refactor: keep config action in default adapter
