+++ config/store.ts
@@ -1,5 +1,4 @@
 const fs = require('fs');
-const { exec } = require('child_process');
 const path = require('path');
 
 type ConfigLeaf = string | number | boolean | null;
@@ -145,20 +144,7 @@ const createConfigStore = ({
     }
   };
 
-  const configAction = (request?: string, keys?: string, value?: string, other?: string[]) => {
-    if (request === 'reset') return resetConfig();
-    // Send full config when no explicit request is provided.
-    if (request === 'get' || !request) return getConfig(keys, value);
-    if (request === 'set') return setConfig(keys, value, other);
-    if (process.env.NODE_ENV !== 'test') {
-      // exec() is async, so actionErr is returned before the help output is printed.
-      exec('ballin', (error: Error | null, stdout: string) => console.log(stdout)); // eslint-disable-line no-console
-    }
-    return configMessages.actionErr;
-  };
-
   return {
-    configAction,
     configPath,
     fetchConfig,
     getConfig,
